rename to AuthFieldSchema
